Moved PackageVersion
- Moved PackageVersion
- Fixed some doc stuff
